+++ sparse/src/KokkosSparse_spmv_handle.hpp
@@ -15,9 +15,13 @@
 //@HEADER
 
 #include <Kokkos_Core.hpp>
+// Use TPL utilities for safely finalizing matrix descriptors, etc.
+#include "KokkosSparse_Utils_cusparse.hpp"
+#include "KokkosSparse_Utils_rocsparse.hpp"
+#include "KokkosSparse_Utils_mkl.hpp"
 
-#ifndef KOKKOSSPARSE_SPMV_HPP_
-#define KOKKOSSPARSE_SPMV_HPP_
+#ifndef KOKKOSSPARSE_SPMV_HANDLE_HPP_
+#define KOKKOSSPARSE_SPMV_HANDLE_HPP_
 
 namespace KokkosSparse {
 
@@ -28,15 +32,71 @@ enum SPMVAlgorithm {
   SPMV_MERGE_PATH
 }
 
-template <class AMatrix, class XVector, class YVector>
+namespace Impl
+{
+  struct CuSparse9_SpMV_Data
+  {
+    ~CuSparse9_SpMV_Data()
+    {
+    }
+
+    cusparseMatDescr_t mat;
+    size_t bufferSize = 0;
+    void* buffer = nullptr;
+  };
+
+  // Data used by cuSPARSE 10 and up
+  struct CuSparse_SpMV_Data
+  {
+    ~CuSparse_SpMV_Data()
+    {
+    }
+
+    cusparseSpMatDescr_t mat;
+  };
+
+  struct RocSparse_SpMV_Data
+  {
+    ~RocSparse_SpMV_Data()
+    {
+    }
+
+    rocsparse_mat_descr mat;
+    rocsparse_spmat_descr spmat;
+    size_t bufferSize = 0;
+    void* buffer = nullptr;
+  };
+
+  struct MKL_SpMV_Data
+  {
+    sparse_matrix_t mat;
+    matrix_descr descr;
+  };
+
+#if defined(KOKKOS_ENABLE_SYCL) && \
+  !defined(KOKKOSKERNELS_ENABLE_TPL_MKL_SYCL_OVERRIDE)
+  struct OneMKL_SpMV_Data
+  {
+  };
+#endif
+}
+
+template <class ExecutionSpace, class AMatrix>
 class SPMVHandle
 {
+public:
   SPMVHandle(SPMVAlgorithm algo_ = SPMV_DEFAULT)
     : algo(algo_)
   {}
 
-  SPMVAlgorithm algo;
+  ~SPMVHandle()
+  {
+  }
 
+  SPMVAlgorithm get_algorithm() const {return algo};
+
+private:
+  SPMVAlgorithm algo;
 };
 
 }
